requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
argh==0.26.2
auth==0.5.3
awscli==1.21.10
awsebcli==3.20.2
baton==1.0.1
blinker==1.4
boto3==1.18.53
botocore==1.21.65
cement==2.8.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.3
cryptography==3.4.8
curlify==2.2.1
DateTime==4.3
defusedxml==0.7.1
django-allauth==0.45.0
django-autocomplete-light==3.8.2
django-baton==2.2.3
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
dnspython==2.1.0
docker-py==1.9.0
docutils==0.15.2
enum34==1.1.10
eventlet==0.32.0
falcon==3.0.1
future==0.16.0
greenlet==1.1.2
gunicorn==20.1.0
hgicommon==1.2.0
hgijson==1.3.1
idna==3.2
inflection==0.5.1
Jinja2==3.0.2
jmespath==0.10.0
keyring==9.0
keyrings.alt==4.1.0
MarkupSafe==2.0.1
mongoengine==0.23.1
mysqlclient==2.0.3
oauthlib==3.1.1
pathspec==0.5.9
pathtools==0.1.2
Pillow==8.3.2
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
PyMySQL==1.0.2
pypiwin32==223
python-dateutil==2.8.2
python3-openid==3.2.0

pywin32==302
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
semantic-version==2.8.5
serializers==0.2.4
six==1.14.0
termcolor==1.1.0
urllib3==1.26.7
watchdog==0.8.3
wcwidth==0.1.9
websocket-client==1.2.1
Werkzeug==2.0.2
zope.interface==5.4.0
